requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.3.0+cu121
torchvision==0.18.0+cu121
torchaudio==2.3.0
cupy-cuda12x
transformers
librosa
noisereduce
numpy<2
pydub
soundfile
